+++ examples/example-next-13/src/app/[locale]/layout.tsx
@@ -1,5 +1,5 @@
-import {Inter} from '@next/font/google';
 import clsx from 'clsx';
+import {Inter} from 'next/font/google';
 import {notFound} from 'next/navigation';
 import {createTranslator, NextIntlClientProvider} from 'next-intl';
 import {ReactNode} from 'react';
@@ -12,8 +12,20 @@ type Props = {
   params: {locale: string};
 };
 
+async function getMessages(locale: string) {
+  try {
+    return (await import(`../../../messages/${locale}.json`)).default;
+  } catch (error) {
+    notFound();
+  }
+}
+
+export async function generateStaticParams() {
+  return ['en', 'de'].map((locale) => ({locale}));
+}
+
 export async function generateMetadata({params: {locale}}: Props) {
-  const messages = (await import(`../../../messages/${locale}.json`)).default;
+  const messages = await getMessages(locale);
 
   // You can use the core (non-React) APIs when you have to use next-intl
   // outside of components. Potentially this will be simplified in the future
@@ -29,12 +41,7 @@ export default async function LocaleLayout({
   children,
   params: {locale}
 }: Props) {
-  let messages;
-  try {
-    messages = (await import(`../../../messages/${locale}.json`)).default;
-  } catch (error) {
-    notFound();
-  }
+  const messages = await getMessages(locale);
 
   return (
     <html className="h-full" lang={locale}>