+++ packages/use-intl/src/core/Formats.tsx
@@ -4,6 +4,7 @@ import NumberFormatOptions from './NumberFormatOptions';
 type Formats = {
   number: Record<string, NumberFormatOptions>;
   dateTime: Record<string, DateTimeFormatOptions>;
+  list: Record<string, Intl.ListFormatOptions>;
 };
 
 export default Formats;