+++ packages/use-intl/CHANGELOG.md
@@ -3,6 +3,80 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 2.15.1 (2023-06-21)
+
+
+### Bug Fixes
+
+* Allow usage of `next-intl/link` and `usePathname` outside of Next.js ([#338](https://github.com/amannn/next-intl/issues/338)) ([6e1a56c](https://github.com/amannn/next-intl/commit/6e1a56c8c7b708b578033b41c31436a4dde32afc)), closes [#337](https://github.com/amannn/next-intl/issues/337)
+
+
+
+
+
+# 2.15.0 (2023-06-20)
+
+
+### Features
+
+* Add `format.list(…)` for formatting conjunctions and disjunctions ([#327](https://github.com/amannn/next-intl/issues/327) by [@stefanprobst](https://github.com/stefanprobst)) ([32cda32](https://github.com/amannn/next-intl/commit/32cda32f47112915bb2032f3f9cc02bf3a4e833b))
+
+
+
+
+
+## 2.14.6 (2023-05-22)
+
+**Note:** Version bump only for package use-intl
+
+
+
+
+
+## 2.14.5 (2023-05-22)
+
+
+### Bug Fixes
+
+* Set `SameSite` attribute for locale cookie to `strict` ([#302](https://github.com/amannn/next-intl/issues/302)) ([0a6bce5](https://github.com/amannn/next-intl/commit/0a6bce5d57733487b99a7da5037c6195b9d2779b)), closes [#301](https://github.com/amannn/next-intl/issues/301)
+
+
+
+
+
+## 2.14.3 (2023-05-22)
+
+
+### Bug Fixes
+
+* Accept `ref` for `next-intl/link` ([#300](https://github.com/amannn/next-intl/issues/300)) ([4d7cc17](https://github.com/amannn/next-intl/commit/4d7cc17de723c23fff81e2d77623f734a7cc9363)), closes [#299](https://github.com/amannn/next-intl/issues/299)
+
+
+
+
+
+## 2.14.2 (2023-05-12)
+
+
+### Bug Fixes
+
+* Fix forwarding of request headers in middleware ([#269](https://github.com/amannn/next-intl/issues/269) by @ARochniak) ([4ecbab5](https://github.com/amannn/next-intl/commit/4ecbab55c53d88a287a11237eea80bd66233f8c1)), closes [#266](https://github.com/amannn/next-intl/issues/266)
+
+
+
+
+
+## 2.14.1 (2023-05-11)
+
+
+### Bug Fixes
+
+* Fix support for older Next.js versions by moving `Link` to `next-intl/link` ([#288](https://github.com/amannn/next-intl/issues/288)) ([f26ef99](https://github.com/amannn/next-intl/commit/f26ef999bf92c142d56d0009259e5a224c5dec5b)), closes [#287](https://github.com/amannn/next-intl/issues/287)
+
+
+
+
+
 # 2.14.0 (2023-05-10)